fix:【运行中工作流】并行或串行任务执行过程中，执行用户的信息不正确

See merge request job_workflow_platform/workflow-designer-ui!557

+++ src/view/workflow/component/projectTable.vue
@@ -54,7 +54,7 @@
 </template>
 
 <script>
-import { workFlowRecord, deleteWorkflow } from '@/api/managerWorkflow'
+import { workFlowRecord, deleteWorkflow, listWorkflowQuote } from '@/api/managerWorkflow'
 import { mapState } from 'vuex'
 
 export default {
@@ -97,10 +97,19 @@ export default {
     },
     async onDeleteRow(row) {
       try {
-        await this.$confirm('删除不可恢复, 请确认是否继续?', '提示', {
+        const quoteRes = await listWorkflowQuote({
+          id: row.id,
+          tenantId: this.tenantId,
+        })
+        if (quoteRes.errorInfo.errorCode) {
+          this.$message.error(quoteRes.errorInfo.errorMsg)
+          return
+        }
+        const confirmMsg = quoteRes.result?.msg ?? '删除不可恢复, 请确认是否继续?'
+        await this.$confirm(confirmMsg, '提示', {
           confirmButtonText: '确定',
           cancelButtonText: '取消',
-          cancelButtonClass: "btn-custom-cancel",
+          cancelButtonClass: 'btn-custom-cancel',
           type: 'warning',
         })
         const { errorInfo } = await deleteWorkflow({
@@ -187,7 +196,6 @@ export default {
 </style>
 
 <style lang="scss">
-
 .btn-custom-cancel {
   @include cancelbutton;
 }
